Fix narrowing conversions and pass drawable lists by const ref
- Add static_cast for narrowing conversions in CharCreation, Voice, TextPainter,
  ListWidget, and WorldLayout (size_t vs int comparisons)
- Fix unsigned/signed comparison in Player and UniverseServer
- Change setPartDrawables/addPartDrawables to take const ref in NetworkedAnimator
- Change drawDrawables to take const ref in CanvasWidget and WidgetLuaBindings
- Fix variable declaration style (separate declarations) in TextPainter
- Use size_t for loop index in Songbook parseABC

+++ source/game/StarWorldLayout.cpp
@@ -947,10 +947,10 @@ void WorldLayout::addLayer(uint64_t seed, int yStart, int yBase, String const& p
   layer.cells.prepend(layer.cells.last());
 
   int yRange = Root::singleton().assets()->json("/world_template.config:playerStartSearchYRange").toInt();
-  int i = 0;
+  size_t i = 0;
   int lastBoundary = 0;
   for (auto const& region : layer.cells) {
-    int nextBoundary = i < static_cast<int>(layer.boundaries.size()) ? layer.boundaries[i] : worldWidth;
+    int nextBoundary = i < layer.boundaries.size() ? layer.boundaries[i] : worldWidth;
     if (spawnBiomeIndexes.contains(region->blockBiomeIndex))
       m_playerStartSearchRegions.append(RectI(lastBoundary, std::max(0, yBase - yRange), nextBoundary, std::min(worldHeight, yBase + yRange)));
     lastBoundary = nextBoundary;
